Implement identical treatment of display fields
The gist of this improvement is that there is a PrintTime call before
displaying each field (side), as opposed to a single time for the entire
display. While it is twice as much as necessary, it ensures there is a
full temporal symmetry between the processing of the two display fields.
(It also reduces the Main Loop to PrintTime + OutputField.)