Delay checking Control button status on boot
The Device Info display can be invoked by booting the device while
pressing the Control button. For this reason, the FW needs to check the
state of this GPIO, whether to enter Device Info mode or not. However,
due to the filter cap (C9) added in HW revision 1.3.1, this check needs
to be delayed, to allow the cap to charge upon power-up (otherwise an
unpressed button is also perceived as pressed).